requirements.txt
check50
submit50

requests
pytest

pylint

lyricsgenius
python-dotenv
SQLAlchemy
matplotlib
